Rename :sampler and :consistent-sampling base packages to avoid conflict
