+++ consistent-sampling/src/main/java/io/opentelemetry/contrib/sampler/consistent/ConsistentComposedOrSampler.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.contrib.samplers;
+package io.opentelemetry.contrib.sampler.consistent;
 
 import static java.util.Objects.requireNonNull;
